Move E2E modals
